working on front-end

+++ src/app/menu/menu.component.ts
@@ -15,17 +15,28 @@ export class MenuComponent {
     @Input() category: Category;
 
     onEdit(menu) {
+        this.toggleInput()
         this.menuService.editMenu(menu)
     }
 
     onDelete(menu) {
-        this.menuService.deleteMenu(menu)
-        .subscribe(
-            result => console.log(result)
-        )
+        if (confirm("Are you sure you really want to delete this menu?")) {
+            this.menuService.deleteMenu(menu)
+                .subscribe(
+                    result => console.log(result)
+                )
+        }
     }
 
-    isLoggedIn() {    
+    toggleInput() {
+        let button = document.getElementById("toggle-button");
+        let form = document.getElementById("form-container");
+        form.style.display = 'block';
+        button.style.paddingBottom = '20px';
+    }
+
+    isLoggedIn() {
         return this.menuService.isLoggedIn();
     }
+    
 }
